WinGui: Initial implementation of the new JSON API for encoding. Not exposed to end users yet.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6654 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ win/CS/HandBrake.ApplicationServices/Services/Encode/EventArgs/EncodeProgressEventArgs.cs
@@ -22,19 +22,19 @@ namespace HandBrake.ApplicationServices.Services.Encode.EventArgs
         /// Gets or sets PercentComplete.
         /// </summary>
         [DataMember]
-        public float PercentComplete { get; set; }
+        public double PercentComplete { get; set; }
 
         /// <summary>
         /// Gets or sets CurrentFrameRate.
         /// </summary>
         [DataMember]
-        public float CurrentFrameRate { get; set; }
+        public double CurrentFrameRate { get; set; }
 
         /// <summary>
         /// Gets or sets AverageFrameRate.
         /// </summary>
         [DataMember]
-        public float AverageFrameRate { get; set; }
+        public double AverageFrameRate { get; set; }
 
         /// <summary>
         /// Gets or sets EstimatedTimeLeft.
